midway through tidyup
