+++ main.js
@@ -19,18 +19,9 @@ const vary = require('./src/middleware/vary');
 const cache = require('./src/middleware/cache');
 const nUi = require('@financial-times/n-ui');
 const verifyAssets = require('./src/lib/verify-assets');
-
+const backendAuthentication = require('./src/middleware/backend-authentication');
 const headCssMiddleware = require('./src/middleware/head-css');
-const backendKeys = [];
-if (process.env.FT_NEXT_BACKEND_KEY) {
-	backendKeys.push(process.env.FT_NEXT_BACKEND_KEY);
-}
-if (process.env.FT_NEXT_BACKEND_KEY_OLD) {
-	backendKeys.push(process.env.FT_NEXT_BACKEND_KEY_OLD);
-}
-if (process.env.FT_NEXT_BACKEND_KEY_OLDEST) {
-	backendKeys.push(process.env.FT_NEXT_BACKEND_KEY_OLDEST);
-}
+const healthChecks = require('./src/lib/health-checks');
 
 module.exports = function(options) {
 
@@ -78,7 +69,6 @@ module.exports = function(options) {
 	app.locals.__environment = process.env.NODE_ENV || '';
 	app.locals.__isProduction = app.locals.__environment.toUpperCase() === 'PRODUCTION';
 	app.locals.__rootDirectory = directory;
-	const healthChecks = options.healthChecks;
 
 	//Remove x-powered-by header
 	app.set('x-powered-by', false);
@@ -89,24 +79,7 @@ module.exports = function(options) {
 
 	// Only allow authorized upstream applications access
 	if (options.withBackendAuthentication) {
-		app.use(function (req, res, next) {
-			// allow static assets through
-			if (req.path.indexOf('/' + name) === 0 ||
-				// allow healthchecks etc. through
-				req.path.indexOf('/__') === 0) {
-				next();
-			} else if (backendKeys.indexOf(req.get('FT-Next-Backend-Key')) !== -1) {
-				res.set('FT-Backend-Authentication', true);
-				next();
-			} else {
-				res.set('FT-Backend-Authentication', false);
-				if (process.env.NODE_ENV === 'production') {
-					res.sendStatus(401);
-				} else {
-					next();
-				}
-			}
-		});
+		app.use(backendAuthentication(name));
 	} else {
 		nLogger.warn({ event: 'BACKEND_AUTHENTICATION_DISABLED', message: 'Backend authentication is disabled, this app is exposed directly to the internet' });
 	}
@@ -120,39 +93,7 @@ module.exports = function(options) {
 		res.sendStatus(418);
 	});
 
-	app.get(/\/__health(?:\.([123]))?$/, function(req, res) {
-		res.set({ 'Cache-Control': 'private, no-cache, max-age=0' });
-		const checks = healthChecks.map(function(check) {
-			return check.getStatus();
-		});
-		if (checks.length === 0) {
-			checks.push({
-				name: 'App has no healthchecks',
-				ok: false,
-				severity: 3,
-				businessImpact: 'If this application encounters any problems, nobody will be alerted and it probably will not get fixed.',
-				technicalSummary: 'This app has no healthchecks set up',
-				panicGuide: 'Don\'t Panic',
-				lastUpdated: new Date()
-			});
-		}
-		if (req.params[0]) {
-			checks.forEach(function(check) {
-				if (check.severity <= Number(req.params[0]) && check.ok === false) {
-					res.status(500);
-				}
-			});
-		}
-
-		res.set('Content-Type', 'application/json');
-		res.send(JSON.stringify({
-			schemaVersion: 1,
-			name: `Next FT.com ${app.locals.__name} in ${process.env.REGION || 'unknown region'}`,
-			systemCode: options.systemCode,
-			description: description,
-			checks: checks
-		}, undefined, 2));
-	});
+	healthChecks(app, options, description);
 
 
 	let handlebarsPromise = Promise.resolve();
@@ -176,7 +117,7 @@ module.exports = function(options) {
 	app.use(cache);
 	app.use(vary);
 
-	verifyAssets.verify(app.locals);
+
 
 	metrics.init({ app: name, flushEvery: 40000 });
 	app.use(function(req, res, next) {
@@ -186,6 +127,7 @@ module.exports = function(options) {
 	});
 
 	serviceMetrics.init(options.serviceDependencies);
+
 
 	app.get('/__about', function(req, res) {
 		res.set({ 'Cache-Control': 'no-cache' });
@@ -199,12 +141,20 @@ module.exports = function(options) {
 		app.use(flags.middleware);
 	}
 
+
+	verifyAssets.verify(app.locals);
+
 	if (options.hasNUiBundle) {
 		if (!options.withFlags) {
 			throw new Error('To use n-ui bundle please also enable flags by passing in `withFlags: true` as an option to n-express');
 		}
 		app.use(nUi.middleware);
 	}
+
+	// get head css
+	const readFile = denodeify(require('fs').readFile);
+	const headCssPromise = options.hasHeadCss ? readFile(directory + '/public/head.css', 'utf-8') : Promise.resolve();
+	app.use(headCssMiddleware(headCssPromise));
 
 	if (options.withAnonMiddleware) {
 		app.use(anon.middleware);
@@ -214,11 +164,6 @@ module.exports = function(options) {
 		flagsPromise.then(navigation.init);
 		app.use(navigation.middleware);
 	}
-
-	// get head css
-	const readFile = denodeify(require('fs').readFile);
-	const headCssPromise = options.hasHeadCss ? readFile(directory + '/public/head.css', 'utf-8') : Promise.resolve();
-	app.use(headCssMiddleware(headCssPromise));
 
 	if (options.withHandlebars) {
 		app.use(function (req, res, next) {
